Refactors countTotalAndAverage. Passes 'avg' to App as a prop, hands this down to Overview
