+++ fec-client/src/App.jsx
@@ -73,6 +73,7 @@ height: 25%;
 
 let App = () => {
   let [productId, setProductId] = useState(65631);
+  let [avg, setAvg] = useState(0);
 
   //save product id on local storage to access it anywhere
   //window.localStorage.getItem("ProductId") -> return '65631'
@@ -91,7 +92,7 @@ let App = () => {
       </H1Div>
       <div>
         <OverviewDiv>
-          <Overview productId={productId} setProductId={setProductId}/>
+          <Overview productId={productId} setProductId={setProductId} avg = {avg}/>
         </OverviewDiv>
 
         <RelatedDiv>
@@ -103,7 +104,7 @@ let App = () => {
         </QandADiv>
 
         <RatingsDiv>
-          <Ratings productId={productId}/>
+          <Ratings productId={productId} setAvg = {setAvg}/>
         </RatingsDiv>
       </div>
     </AppDiv>
