+++ fec-client/src/components/Ratings/index.jsx
@@ -5,13 +5,13 @@ import axios from 'axios';
 import ReviewList from './ReviewList';
 import RatingsBreakdown from './RatingsBreakdown';
 import Write from './Write';
+import countTotalAndAverage from './countTotalAndAverage.js';
 
 const RatingsContainer = styled.div`
 font-weight: 300;
 `;
 
-
-const Ratings = ({productId}) => {
+const Ratings = ({productId, setAvg}) => {
   const [data, setData] = useState([]);
   const [currentData, setCurrentData] = useState([]);
   const [metaData, setMetaData] = useState({});
@@ -30,6 +30,7 @@ const Ratings = ({productId}) => {
     axios.get('/api', {headers: {path: `/reviews/meta?product_id=${productId}`}})
       .then((response) => {
         setMetaData(response.data);
+        setAvg(countTotalAndAverage(response.data.ratings)[1]);
       })
       .catch((err) => console.log(err));
   }, []);
